// ibov-frontend-reactjs/src/interfaces/Fii.ts
export interface IFii {
  companyId?: number;
  companyName?: string;
  ticker?: string;
  price?: number;
  gestao?: string;
  dy?: number;
  p_vp?: number;
  valorpatrimonialcota?: number;
  liquidezmediadiaria?: number;
  percentualcaixa?: number;
  dividend_cagr?: number;
  cota_cagr?: number;
  numerocotistas?: number;
  numerocotas?: number;
  patrimonio?: number;
  lastdividend?: number;
}